refactor: split `fakes` module into dedicated `ansible` and `fs` modules for improved organization.
